# apps/auth-manager/openapi3.yaml
openapi: 3.0.1
info:
  title: auth-manager
  description: API server for managing all the authentication data
  version: 1.12.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
...
  '/client/{clientName}/connection':
    parameters:
      - $ref: '#/components/parameters/clientParam'
    get:
      operationId: getClientConnections
      summary: gets the connections for a specific client
      tags:
        - connection
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/connection'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  '/client/{clientName}/connection/{environment}':
    parameters:
      - $ref: '#/components/parameters/clientParam'
      - $ref: '#/components/parameters/environmentPathParam'
    get:
      operationId: getClientEnvironmentConnections
      summary: get the latest client connection for specific environment
      tags:
        - connection
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/connection'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  '/client/{clientName}/connection/{environment}/{version}':
    parameters:
      - $ref: '#/components/parameters/clientParam'
      - $ref: '#/components/parameters/environmentPathParam'
      - $ref: '#/components/parameters/versionParam'
    get:
      operationId: getClientVersionedConnection
      summary: get a specfic client connection for specific environment
      tags:
        - connection
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connection'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  '/client/{clientName}/connection/{environment}/latest':
    parameters:
      - $ref: '#/components/parameters/clientParam'
      - $ref: '#/components/parameters/environmentPathParam'
    get:
      operationId: getClientLatestConnection
      summary: get the latest client connection for specific environment
      tags:
        - connection
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connection'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /connection:
    get:
      operationId: getConnections
      summary: get a connections based on filters
      tags:
        - connection
      parameters:
        - in: query
          name: name
          description: search by client name (partial match, case-insensitive)
          schema:
            type: string
        - in: query
          name: onlyLatest
          description: if true, returns only the latest version per (name, environment) pair
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/environmentQueryParam'
        - in: query
          name: isEnabled
          schema:
            type: boolean
        - in: query
          name: isNoBrowser
          schema:
            type: boolean
        - in: query
          name: isNoOrigin
          schema:
            type: boolean
        - in: query
          name: domains
          schema:
            type: array
            uniqueItems: true
            items:
              type: string
        - in: query
          name: sort
          description: >-
            Sorts the results based on the value of one or more properties.

            The value is a comma-separated list of property names and sort order.

            properties should be separated by a colon and sort order should be either asc or desc. For example: created-at:asc,name:asc

            The default sort order is ascending. If the sort order is not specified, the default sort order is used. Each property is only allowed to appear once in the list.

            The available properties are:

            - created-at

            - name

            - version

            - enabled

            - environment
          example:
            - created-at:asc
            - name:asc
          required: false
          explode: false
          schema:
            default:
              - name:asc
              - environment:asc
              - version:desc
            type: array
            uniqueItems: true
            items:
              example: name:asc
              type: string
              pattern: ^(created-at|name|version|enabled|environment)(:asc|:desc){0,1}$
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/paginationResponse'
                  - type: object
                    required:
                      - items
                    properties:
                      items:
                        description: list of clients
                        type: array
                        uniqueItems: true
                        minItems: 0
                        maxItems: 1000
                        items:
                          $ref: '#/components/schemas/connection'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    post:
      operationId: upsertConnection
      summary: creates a new connection or updates it based on the version
      tags:
        - connection
      parameters:
        - in: query
          name: shouldIgnoreTokenErrors
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/connection'
      responses:
        '200':
          description: updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connection'
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connection'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        '409':
          $ref: '#/components/responses/409Conflict'
        '500':
          $ref: '#/components/responses/500InternalServerError'
...
  /bundle:
    get:
      operationId: getBundles
      summary: get bundles by filter
      tags:
        - bundle
      parameters:
        - in: query
          name: createdBefore
          description: filters all clients created before given date
          schema:
            type: string
            format: date-time
        - in: query
          name: createdAfter
          description: filters all clients created after given date
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/environmentQueryParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/bundle'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  '/bundle/{id}':
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: number
          format: int32
          minimum: 1
    get:
      operationId: getBundle
      summary: get a specific bundle
      tags:
        - bundle
      parameters:
        - $ref: '#/components/parameters/environmentQueryParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bundle'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /domain:
    get:
      parameters:
        - in: query
          name: sort
          description: >-
            Sorts the results based on the value of one or more properties.

            The value is a comma-separated list of property names and sort order.

            properties should be separated by a colon and sort order should be either asc or desc. For example: domain:asc

            The default sort order is ascending. If the sort order is not specified, the default sort order is used. Each property is only allowed to appear once in the list.

            The available properties are:

            - domain
          example:
            - domain:asc
          required: false
          explode: false
          schema:
            default:
              - domain:asc
            type: array
            uniqueItems: true
            items:
              example: domain:asc
              type: string
              pattern: ^(domain)(:asc|:desc){0,1}$
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      operationId: getDomains
      summary: get all the domains
      tags:
        - domain
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/paginationResponse'
                  - type: object
                    required:
                      - items
                    properties:
                      items:
                        description: list of clients
                        type: array
                        uniqueItems: true
                        minItems: 0
                        items:
                          $ref: '#/components/schemas/domain'
        '415':
          $ref: '#/components/responses/415UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/500InternalServerError'
    post:
      operationId: createDomain
      summary: create a new domain
      tags:
        - domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/domain'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/domain'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '409':
          $ref: '#/components/responses/409Conflict'
        '500':
          $ref: '#/components/responses/500InternalServerError'
security: []
components:
  responses:
    400BadRequest:
      description: BadRequest
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    404NotFound:
      description: Not Found - If client does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    409Conflict:
      description: conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    415UnsupportedMediaType:
      description: Unsupported Media Type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    500InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
  parameters:
    page:
      name: page
      in: query
      description: >
        page number for pagination. The value is 1-based, meaning the first page
        is 1.

        If the value is greater than the total number of pages, an empty array will be returned.

        This simplifies the pagination logic on the client side.
      schema:
        type: integer
        format: int32
        default: 1
        minimum: 1
    pageSize:
      name: page_size
      in: query
      description: number of items per page
      schema:
        type: integer
        format: int32
        default: 10
        maximum: 100
        minimum: 1
    clientParam:
      name: clientName
      in: path
      required: true
      schema:
        type: string
        minLength: 3
...
    versionParam:
      name: version
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/version'
    environmentQueryParam:
      name: environment
      in: query
      schema:
        type: array
        uniqueItems: true
        items:
          $ref: '#/components/schemas/environment'
    environmentPathParam:
      name: environment
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/environment'
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    environment:
      type: string
      enum:
        - np
        - stage
        - prod
    version:
      type: number
      default: 1
      format: int32
      minimum: 1
...
    name:
      type: string
      minLength: 2
      maxLength: 32
    domain:
      type: object
      required:
        - name
      properties:
        name:
          $ref: '#/components/schemas/name'
...
    connection:
      type: object
      required:
        - name
        - enabled
        - environment
        - token
        - domains
        - allowNoBrowserConnection
        - allowNoOriginConnection
        - origins
        - version
      properties:
        name:
          $ref: '#/components/schemas/name'
        environment:
          $ref: '#/components/schemas/environment'
        version:
          $ref: '#/components/schemas/version'
        enabled:
          type: boolean
          default: true
        token:
          type: string
        domains:
          type: array
          uniqueItems: true
          items:
            type: string
          minLength: 1
        createdAt:
          type: string
          format: date-time
          readOnly: true
        allowNoBrowserConnection:
          type: boolean
        allowNoOriginConnection:
          type: boolean
        origins:
          type: array
          uniqueItems: true
          items:
            type: string
...
    bundle:
      type: object
      properties:
        id:
          type: number
          format: int32
        hash:
          type: string
        metadata:
          type: object
          additionalProperties: true
        assets:
          type: array
          uniqueItems: true
          items:
            type: object
            required:
              - name
              - version
            properties:
              name:
                type: string
              version:
                $ref: '#/components/schemas/version'
        connections:
          type: array
          uniqueItems: true
          items:
            type: object
            required:
              - name
              - version
            properties:
              name:
                type: string
              version:
                $ref: '#/components/schemas/version'
        environment:
          allOf:
            - $ref: '#/components/schemas/environment'
            - readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        keyVersion:
          $ref: '#/components/schemas/version'
        opaVersion:
          type: string
          description: OPA version used to generate the bundle
    paginationResponse:
      type: object
      required:
        - total
      properties:
        total:
          type: integer
          format: int32
          description: total number of items